[editor canvas] Remove state.js in favour of state/index.js
